Parameters: organize parameters into directory
This commit moves all the parameters and their test cases to make it so
they are not a jumbled mess of files as they were before.
