+++ .gitignore
@@ -35,3 +35,4 @@ tags
 
 coverage/
 
+.#*.ts
