+++ test/Parameter/AltRepParam.spec.ts
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import AltRepParam from "../src/AltRepParam";
+import AltRepParam from "../../src/Parameter/AltRepParam";
 import { expect } from "chai";
 import "mocha";
