+++ test/Parameter/CalUserTypeParam.spec.ts
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import CalUserTypeParam from "../src/CalUserTypeParam";
+import CalUserTypeParam from "../../src/Parameter/CalUserTypeParam";
 import { expect } from "chai";
 import "mocha";
