+++ test/Parameter/DelegateesParam.spec.ts
@@ -15,7 +15,7 @@
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-import DelegateesParam from "../src/DelegateesParam";
+import DelegateesParam from "../../src/Parameter/DelegateesParam";
 import { expect } from "chai";
 import "mocha";
