+++ test/Parameter/DelegatorsParam.spec.ts
@@ -15,7 +15,7 @@
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-import DelegatorsParam from "../src/DelegatorsParam";
+import DelegatorsParam from "../../src/Parameter/DelegatorsParam";
 import { expect } from "chai";
 import "mocha";
